Create Const File and Move Version There
