+++ .bumpversion.cfg
@@ -5,7 +5,7 @@ tag = False
 
 [bumpversion:file:setup.py]
 
-[bumpversion:file:insteon_mqtt/version.py]
+[bumpversion:file:insteon_mqtt/const.py]
 
 [bumpversion:file:README.md]
 search = Version: {current_version}
