+++ insteon_mqtt/const.py
@@ -1,0 +1,16 @@
+#===========================================================================
+#
+# Insteon-MQTT Constants File
+#
+#===========================================================================
+""" Constants File
+
+This file can be used to store and import constant values throughout the
+code.  This was added after the code base had matured, so it is largely
+unused.  The primary impetus for this is to allow importing of the version
+variable throughout the code without causing a cyclic import
+"""
+
+__version__ = "0.8.3"
+
+#===========================================================================
