--- insteon_mqtt/version.py
@@ -1,14 +1,0 @@
-#===========================================================================
-#
-# Insteon-MQTT Verson File
-#
-#===========================================================================
-""" Version File
-
-This is used to allow importing of the version variable throughout the code
-without causing a cyclic import
-"""
-
-__version__ = "0.8.3"
-
-#===========================================================================
